+++ src/components/ReadPage/PostComment.js
@@ -6,6 +6,7 @@ import { useParams } from "react-router-dom";
 import { useSelector } from "react-redux";
 import Axios from "axios";
 import CommentEdit from "./CommentEdit";
+import CommentList from "./CommentList";
 import { style } from "@mui/system";
 
 //게시물 작성 완료 페이지의 댓글창
@@ -18,20 +19,6 @@ const PostComment = ({ commentData }) => {
   // const postListShow = useSelector(state => state.post.postListShow);
   // const [postList, setPostList] = useState([]);
   const [message, setMessage] = useState("");
-  const [editOpen, setEditOpen] = useState(false);
-  const [inputEditHandle, setInputEditHandle] = useState("");
-
-  const inputHandle = e => {
-    setInputEditHandle(e.target.value);
-  };
-
-  const contentInputHandle = () => {
-    setEditOpen(true);
-  };
-
-  const contentInputCancel = () => {
-    setEditOpen(false);
-  };
 
   const handleMessage = event => {
     setMessage(event.target.value);
@@ -98,70 +85,11 @@ const PostComment = ({ commentData }) => {
           ? commentData.map((comment, index) => {
               return (
                 <>
-                  <Section style={{ paddingTop: "15px" }}>
-                    <div
-                      style={{
-                        display: "flex",
-                        alignItems: "center",
-                      }}
-                      key={index}
-                    >
-                      {comment.commenterProfile ? (
-                        <img
-                          src={
-                            "data:image/jpeg;base64," +
-                            comment.commenterProfile?.imageData
-                          }
-                          alt="사용자 이미지"
-                        />
-                      ) : (
-                        <img
-                          src="https://cdn.pixabay.com/photo/2015/10/05/22/37/blank-profile-picture-973460_1280.png"
-                          alt="사용자 이미지"
-                        />
-                      )}
-                      <div>
-                        <div
-                          style={{ display: "flex", flexDirection: "column" }}
-                        >
-                          <div className="Name">
-                            {comment.commenterNickname}
-                          </div>
-                          <div className="Time">
-                            {comment.createdTime.slice(0, 10)}
-                            &nbsp;
-                            {comment.createdTime.slice(11, 16)}
-                          </div>
-                        </div>
-                      </div>
-                    </div>
-                    {memberId == comment.commenterId && (
-                      <CommentEdit
-                        contentInputHandle={contentInputHandle}
-                        comment={comment}
-                      />
-                    )}
-                  </Section>
-                  <Comment>
-                    <p>{comment.content}</p>
-                    {/* {수정 버튼 누르고 난 뒤} */}
-                    {editOpen && (
-                      <div>
-                        <input
-                          className="input"
-                          type="text"
-                          placeholder={comment.content}
-                          value={inputEditHandle}
-                          name="contentInput"
-                          onChange={inputHandle}
-                        />
-                        <div className="editButtonHandle">
-                          <button onClick={contentInputCancel}>취소</button>
-                          <button className="complete">완료</button>
-                        </div>
-                      </div>
-                    )}
-                  </Comment>
+                  <CommentList
+                    comment={comment}
+                    index={index}
+                    memberId={memberId}
+                  />
                 </>
               );
             })
@@ -170,70 +98,5 @@ const PostComment = ({ commentData }) => {
     </div>
   );
 };
-
-const Section = styled.div`
-  display: flex;
-  justify-content: space-between;
-  margin-bottom: 18px;
-
-  img {
-    display: block;
-    width: 52px;
-    height: 52px;
-    margin-right: 16px;
-    border-radius: 50%;
-    object-fit: cover;
-  }
-
-  .Name {
-    color: #333;
-    font-weight: 700;
-  }
-
-  .Time {
-    font-size: 14px;
-    line-height: 126.5%;
-    letter-spacing: -0.005em;
-    color: #9f9f9f;
-  }
-`;
-
-const Comment = styled.div`
-  input {
-    width: 100%;
-    border-radius: 8px;
-    padding: 8px 12px;
-    border: 1px solid #e1e1e1;
-    outline: none;
-  }
-
-  .editButtonHandle {
-    display: flex;
-    justify-content: start;
-    align-items: center;
-
-    button {
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      background: #e9ecef;
-      padding: 13px;
-      cursor: pointer;
-      border-radius: 4px;
-      width: 60px;
-      height: 1.5rem;
-      margin-bottom: 10px;
-      margin-top: 10px;
-      margin-right: 10px;
-      outline: none;
-      border: 0px solid #ccc;
-      font-weight: bold;
-    }
-    .complete {
-      background-color: black;
-      color: white;
-    }
-  }
-`;
 
 export default PostComment;
